+++ examples/uniswap-read/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @layerzerolabs/uniswap-read-example
 
+## 0.2.5
+
+### Patch Changes
+
+- af91805: Bump to lz-definitions 3.0.59+
+
+## 0.2.4
+
+### Patch Changes
+
+- ce03876: Get latest EndpointIds by bumping lz-definitions
+
 ## 0.2.3
 
 ### Patch Changes