+++ packages/create-lz-oapp/CHANGELOG.md
@@ -1,5 +1,17 @@
 # create-lz-oapp
 
+## 0.5.0
+
+### Minor Changes
+
+- e60d40b: Added lzapp migration example
+
+## 0.4.2
+
+### Patch Changes
+
+- 186442a: add "test:jest" script
+
 ## 0.4.1
 
 ### Patch Changes