+++ packages/devtools-evm-hardhat/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @layerzerolabs/devtools-evm-hardhat
 
+## 2.0.6
+
+### Patch Changes
+
+- af91805: Bump to lz-definitions 3.0.59+
+- 186442a: add "test:jest" script
+- Updated dependencies [af91805]
+- Updated dependencies [186442a]
+  - @layerzerolabs/devtools-evm@1.0.4
+  - @layerzerolabs/devtools@0.4.5
+  - @layerzerolabs/export-deployments@0.0.15
+  - @layerzerolabs/io-devtools@0.1.15
+
+## 2.0.5
+
+### Patch Changes
+
+- ce03876: Get latest EndpointIds by bumping lz-definitions
+- Updated dependencies [ce03876]
+  - @layerzerolabs/devtools-evm@1.0.3
+  - @layerzerolabs/devtools@0.4.4
+
 ## 2.0.4
 
 ### Patch Changes