+++ packages/devtools-ton/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @layerzerolabs/devtools-ton
 
+## 0.2.2
+
+### Patch Changes
+
+- af91805: Bump to lz-definitions 3.0.59+
+- Updated dependencies [af91805]
+- Updated dependencies [186442a]
+  - @layerzerolabs/devtools@0.4.5
+  - @layerzerolabs/io-devtools@0.1.15
+
+## 0.2.1
+
+### Patch Changes
+
+- ce03876: Get latest EndpointIds by bumping lz-definitions
+- Updated dependencies [ce03876]
+  - @layerzerolabs/devtools@0.4.4
+
 ## 0.2.0
 
 ### Minor Changes