+++ packages/protocol-devtools-solana/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @layerzerolabs/protocol-devtools-solana
 
+## 4.0.7
+
+### Patch Changes
+
+- af91805: Bump to lz-definitions 3.0.59+
+- Updated dependencies [af91805]
+- Updated dependencies [186442a]
+  - @layerzerolabs/protocol-devtools@1.1.4
+  - @layerzerolabs/devtools-solana@1.0.6
+  - @layerzerolabs/ua-devtools@3.0.4
+  - @layerzerolabs/devtools@0.4.5
+  - @layerzerolabs/io-devtools@0.1.15
+
+## 4.0.6
+
+### Patch Changes
+
+- ce03876: Get latest EndpointIds by bumping lz-definitions
+- Updated dependencies [ce03876]
+  - @layerzerolabs/protocol-devtools@1.1.3
+  - @layerzerolabs/devtools-solana@1.0.5
+  - @layerzerolabs/ua-devtools@3.0.3
+  - @layerzerolabs/devtools@0.4.4
+
 ## 4.0.5
 
 ### Patch Changes